requirements.txt
gradio
torch
torchvision
opencv-python
numpy
Pillow
fastapi
reportlab
PyPDF2